+++ src/lib.rs
@@ -3,8 +3,12 @@
 #![warn(clippy::pedantic)]
 #![allow(clippy::double_must_use)]
 mod n_choose_k;
-pub mod bit_set;
+pub mod bit_set_array;
 pub mod bit_set_n;
+pub mod bit_set_trait;
 
-pub use bit_set::*;
+
+pub type SetElement = u32;
+pub use bit_set_array::*;
 pub use bit_set_n::*;
+